Finishing basic functionality

+++ Data/Order.cs
@@ -2,8 +2,8 @@ namespace OnlineShop.Data
 {
     public class Order
     {
-        public int Id { get; set;}
+        public long Id { get; set;}
         public string CustomerName {get; set;}
-        public float TotalPrice {get; set;}
+        public decimal TotalPrice {get; set;}
     }
 }
