+++ Data/OrderLineDto.cs
@@ -2,8 +2,8 @@ namespace OnlineShop.Data
 {
     public class OrderLineDto
     {
-        public int Id { get; set;}
-        public int OrderId { get; set;}
+        public long Id { get; set;}
+        public long OrderId { get; set;}
         public int ProductId {get; set;}
         public int Quantity {get; set;}
         public OrderDto Order { get; set; }
